+++ Changelog.md
@@ -6,15 +6,15 @@ This is the changelog for my changes in DesktopPet-Z only. See "Vanilla Changelo
 
 
 
-###### v0.0 (FILL THIS OUT)
+### v0.0 (FILL THIS OUT)
 
-###### v0.1 (FILL THIS OUT)
+### v0.1 (FILL THIS OUT)
 
-###### v0.2 (FILL THIS OUT)
+### v0.2 (FILL THIS OUT)
 
 
 
-###### v0.3 (FILL THIS OUT)
+### v0.3 (FILL THIS OUT)
 
 * (WIP) Flinging
 
@@ -23,4 +23,22 @@ This is the changelog for my changes in DesktopPet-Z only. See "Vanilla Changelo
 v0.3.1 (WIP)
 
 * Made it so that key-animations can play sound effects
+
+
+
+### V0.4 (WIP)
+
+* Pets can now detect the side borders of windows! This also means they can play certain animations when colliding with other pets!
+
+Technical info:
+
+* Pets intentionally phase through windows if they don't have enough room to get past the screen borders
+
+WIP info:
+
+* Pets currently can only detect the right border of windows
+* Pets currently cannot detect right/left window borders when being tossed
+* Pets currently only play NONE-tagged border animations when hitting windows
+
+